+++ artista.h
@@ -1,23 +1,23 @@
 #ifndef ARTISTA_H
 #define ARTISTA_H
-
+#include <cstdint>
 #include <string>
 
 using namespace std;
 
 class Artista{
 private:
-    int age;
+    uint8_t age;
     string country;
     size_t followers;
-    int position;
+    unsigned int position;
 public:
     string name;
-    int idArtist;
+    unsigned int idArtist;
 
     // Constructor
     Artista();
-    Artista(string _name, int _id, int _age, string _country, size_t _followers, int _position);
+    Artista(string _name, unsigned int _id, uint8_t _age, string _country, size_t _followers, unsigned int _position);
 };
 
 #endif // ARTISTA_H
